+++ cirq-core/cirq/experiments/__init__.py
@@ -71,8 +71,3 @@ from cirq.experiments.two_qubit_xeb import (
     parallel_two_qubit_xeb,
     run_rb_and_xeb,
 )
-
-from cirq.experiments.measure_in_random_bases import (
-    append_randomized_measurements,
-    RandomizedMeasurements,
-)
